backend/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Base de datos
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Vector DB (pgvector)
pgvector==0.3.6

# MinIO
minio==7.2.9

# Procesamiento PDF
PyPDF2==3.0.1
pdfplumber==0.11.4
pdf2image==1.17.0
Pillow==11.0.0

# IA y Embeddings
openai==1.54.3
anthropic==0.39.0
langchain==0.3.7
langchain-openai==0.2.8
sentence-transformers==3.2.1

# Utilidades
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
